+++ applications/main/subghz/helpers/subghz_custom_event.h
@@ -24,7 +24,7 @@ typedef enum {
     SubGhzCustomEventSceneShowOnlyRX,
     SubGhzCustomEventSceneAnalyzerLock,
     SubGhzCustomEventSceneAnalyzerUnlock,
-    SubGhzCustomEventSceneAnalyzerMissing,
+    SubGhzCustomEventSceneFeaturePluginMissing,
     SubGhzCustomEventSceneSettingLock,
     SubGhzCustomEventSceneSettingResetToDefault,